Fix STLLaVA garbage output: restore discarded vision tower weights and fix T4 float16 overflow

+++ backend/utils/transformers_patch.py
@@ -57,7 +57,7 @@ def patch_model_loading_kwargs():
                 kwargs["quantization_config"] = BitsAndBytesConfig(
                     load_in_8bit=load_in_8bit,
                     load_in_4bit=load_in_4bit,
-                    bnb_4bit_compute_dtype=torch.float16,
+                    bnb_4bit_compute_dtype=torch.float32,
                     bnb_4bit_use_double_quant=True,
                     bnb_4bit_quant_type="nf4"
                 )
